requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.32
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
websockets==12.0
aiofiles==23.2.1
jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
email-validator
boto3
bcrypt==3.2.2